% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_faq.R
\name{nsgen}
\alias{nsgen}
\title{根据FAQ标准问题及同义记事自动做问题泛化}
\usage{
nsgen(file = "data-raw/JBLH_upload_V1.8.xlsx", skip = 1)
}
\arguments{
\item{file}{文件名称}

\item{skip}{跳过多少行}
}
\value{
返回值
}
\description{
根据FAQ标准问题及同义记事自动做问题泛化
}
\examples{
nsgen()
}
