% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_cartype.R
\name{nsgen_cartype}
\alias{nsgen_cartype}
\title{处理相当问题}
\usage{

  nsgen_cartype(file = "data-raw/捷豹路虎品牌车型知识点模板V2.2_发现运动_需泛化相似问.xlsx",
  sheet_name = "发运-固定内容", new_carType = "极光揽胜")
}
\arguments{
\item{file}{文件名}

\item{sheet_name}{页签}

\item{new_carType}{新车型}
}
\value{
返回值
}
\description{
处理相当问题
}
\examples{
nsgen_cartype()
}
